Setup proper CSRF for AJAX
There is now a view whose purpose is to put the CSRF token in the
response cookies, /csrf/. The Javascript library GETs that URL to make
sure that the cookie is properly set before making any other AJAX calls.

The code to setup the AJAX information is now in a function that gets
called only after the cookie is properly present.